requirements.txt
Django==3.1.2
django-http-exceptions==1.4.0
djangorestframework==3.12.1

